app/requirements.txt
ipython
openai
scikit-learn
pandas
flask
streamlit
elasticsearch
openpyxl
tqdm
notebook==7.1.2
ipywidgets
sentence-transformers
openai-whisper
SpeechRecognition
pydub
gTTS
streamlit-audiorec
psycopg2-binary
pgcli